chore(expr): move unrelated functions out of DataType
