+++ src/query/expression/src/kernels/scatter.rs
@@ -39,6 +39,7 @@ use crate::ColumnBuilder;
 use crate::DataBlock;
 use crate::Scalar;
 use crate::TypeDeserializer;
+use crate::TypeDeserializerImpl;
 use crate::Value;
 
 impl DataBlock {
@@ -194,10 +195,7 @@ impl Column {
                 let mut offsets = Vec::with_capacity(length + 1);
                 offsets.push(0);
                 let builder = ColumnBuilder::from_column(
-                    column
-                        .values
-                        .data_type()
-                        .create_deserializer(length)
+                    TypeDeserializerImpl::with_capacity(&column.values.data_type(), length)
                         .finish_to_column(),
                 );
                 let builder = ArrayColumnBuilder { builder, offsets };
