+++ src/query/expression/src/types.rs
@@ -32,11 +32,8 @@ pub mod variant;
 use std::fmt::Debug;
 use std::ops::Range;
 
-use common_arrow::arrow::bitmap::MutableBitmap;
 use common_arrow::arrow::trusted_len::TrustedLen;
 use enum_as_inner::EnumAsInner;
-use ethnum::i256;
-use ordered_float::OrderedFloat;
 use serde::Deserialize;
 use serde::Serialize;
 
@@ -51,28 +48,16 @@ pub use self::generic::GenericType;
 pub use self::map::MapType;
 pub use self::null::NullType;
 pub use self::nullable::NullableType;
-use self::number::NumberScalar;
 pub use self::number::*;
 pub use self::number_class::*;
-use self::string::StringColumnBuilder;
 pub use self::string::StringType;
 pub use self::timestamp::TimestampType;
 pub use self::variant::VariantType;
-use crate::deserializations::ArrayDeserializer;
-use crate::deserializations::DateDeserializer;
-use crate::deserializations::DecimalDeserializer;
-use crate::deserializations::MapDeserializer;
-use crate::deserializations::NullableDeserializer;
-use crate::deserializations::NumberDeserializer;
-use crate::deserializations::TimestampDeserializer;
-use crate::deserializations::TupleDeserializer;
-use crate::deserializations::VariantDeserializer;
 use crate::property::Domain;
 use crate::values::Column;
 use crate::values::Scalar;
 use crate::ColumnBuilder;
 use crate::ScalarRef;
-use crate::TypeDeserializerImpl;
 
 pub type GenericMap = [DataType];
 
@@ -213,63 +198,6 @@ impl DataType {
             )),
         }
     }
-    pub fn create_deserializer(&self, capacity: usize) -> TypeDeserializerImpl {
-        match self {
-            DataType::Null => 0.into(),
-            DataType::Boolean => MutableBitmap::with_capacity(capacity).into(),
-            DataType::String => StringColumnBuilder::with_capacity(capacity, capacity * 4).into(),
-            DataType::Number(num_ty) => match num_ty {
-                NumberDataType::UInt8 => {
-                    NumberDeserializer::<u8, u8>::with_capacity(capacity).into()
-                }
-                NumberDataType::UInt16 => {
-                    NumberDeserializer::<u16, u16>::with_capacity(capacity).into()
-                }
-                NumberDataType::UInt32 => {
-                    NumberDeserializer::<u32, u32>::with_capacity(capacity).into()
-                }
-                NumberDataType::UInt64 => {
-                    NumberDeserializer::<u64, u64>::with_capacity(capacity).into()
-                }
-                NumberDataType::Int8 => {
-                    NumberDeserializer::<i8, i8>::with_capacity(capacity).into()
-                }
-                NumberDataType::Int16 => {
-                    NumberDeserializer::<i16, i16>::with_capacity(capacity).into()
-                }
-                NumberDataType::Int32 => {
-                    NumberDeserializer::<i32, i32>::with_capacity(capacity).into()
-                }
-                NumberDataType::Int64 => {
-                    NumberDeserializer::<i64, i64>::with_capacity(capacity).into()
-                }
-                NumberDataType::Float32 => {
-                    NumberDeserializer::<F32, f32>::with_capacity(capacity).into()
-                }
-                NumberDataType::Float64 => {
-                    NumberDeserializer::<F64, f64>::with_capacity(capacity).into()
-                }
-            },
-            DataType::Date => DateDeserializer::with_capacity(capacity).into(),
-            DataType::Timestamp => TimestampDeserializer::with_capacity(capacity).into(),
-            DataType::Nullable(inner_ty) => {
-                NullableDeserializer::with_capacity(capacity, inner_ty.as_ref()).into()
-            }
-            DataType::Variant => VariantDeserializer::with_capacity(capacity).into(),
-            DataType::Array(ty) => ArrayDeserializer::with_capacity(capacity, ty).into(),
-            DataType::Map(ty) => MapDeserializer::with_capacity(capacity, ty).into(),
-            DataType::Tuple(types) => TupleDeserializer::with_capacity(capacity, types).into(),
-            DataType::Decimal(types) => match types {
-                DecimalDataType::Decimal128(_) => {
-                    DecimalDeserializer::<i128>::with_capacity(types, capacity).into()
-                }
-                DecimalDataType::Decimal256(_) => {
-                    DecimalDeserializer::<i256>::with_capacity(types, capacity).into()
-                }
-            },
-            _ => unimplemented!(),
-        }
-    }
 
     // Nullable will be displayed as Nullable(T)
     pub fn wrapped_display(&self) -> String {
@@ -296,48 +224,6 @@ impl DataType {
             DataType::String => "VARCHAR".to_string(),
             DataType::Nullable(inner_ty) => format!("{} NULL", inner_ty.sql_name()),
             _ => self.to_string().to_uppercase(),
-        }
-    }
-
-    pub fn default_value(&self) -> Scalar {
-        match self {
-            DataType::Null => Scalar::Null,
-            DataType::EmptyArray => Scalar::EmptyArray,
-            DataType::EmptyMap => Scalar::EmptyMap,
-            DataType::Boolean => Scalar::Boolean(false),
-            DataType::String => Scalar::String(vec![]),
-            DataType::Number(num_ty) => Scalar::Number(match num_ty {
-                NumberDataType::UInt8 => NumberScalar::UInt8(0),
-                NumberDataType::UInt16 => NumberScalar::UInt16(0),
-                NumberDataType::UInt32 => NumberScalar::UInt32(0),
-                NumberDataType::UInt64 => NumberScalar::UInt64(0),
-                NumberDataType::Int8 => NumberScalar::Int8(0),
-                NumberDataType::Int16 => NumberScalar::Int16(0),
-                NumberDataType::Int32 => NumberScalar::Int32(0),
-                NumberDataType::Int64 => NumberScalar::Int64(0),
-                NumberDataType::Float32 => NumberScalar::Float32(OrderedFloat(0.0)),
-                NumberDataType::Float64 => NumberScalar::Float64(OrderedFloat(0.0)),
-            }),
-            DataType::Decimal(ty) => Scalar::Decimal(ty.default_scalar()),
-            DataType::Timestamp => Scalar::Timestamp(0),
-            DataType::Date => Scalar::Date(0),
-            DataType::Nullable(_) => Scalar::Null,
-            DataType::Array(ty) => {
-                let builder = ColumnBuilder::with_capacity(ty, 0);
-                let col = builder.build();
-                Scalar::Array(col)
-            }
-            DataType::Map(ty) => {
-                let builder = ColumnBuilder::with_capacity(ty, 0);
-                let col = builder.build();
-                Scalar::Map(col)
-            }
-            DataType::Tuple(tys) => {
-                Scalar::Tuple(tys.iter().map(|ty| ty.default_value()).collect())
-            }
-            DataType::Variant => Scalar::Variant(vec![]),
-
-            _ => unimplemented!(),
         }
     }
 }
