+++ src/query/service/src/pipelines/processors/transforms/group_by/aggregator_groups_builder.rs
@@ -20,6 +20,7 @@ use common_expression::types::DataType;
 use common_expression::Column;
 use common_expression::HashMethodFixedKeys;
 use common_expression::TypeDeserializer;
+use common_expression::TypeDeserializerImpl;
 use common_io::prelude::FormatSettings;
 
 use crate::pipelines::processors::AggregatorParams;
@@ -125,7 +126,7 @@ impl<'a> GroupColumnsBuilder for SerializedKeysGroupColumnsBuilder<'a> {
         let mut res = Vec::with_capacity(self.group_data_types.len());
         let format = FormatSettings::default();
         for data_type in self.group_data_types.iter() {
-            let mut deserializer = data_type.create_deserializer(rows);
+            let mut deserializer = TypeDeserializerImpl::with_capacity(data_type, rows);
 
             for (_, key) in keys.iter_mut().enumerate() {
                 deserializer.de_binary(key, &format)?;
